float64 instead of 32 for time and prices

+++ src/qrm_core/engine.py
@@ -37,9 +37,9 @@ def simulate_QRM_jit(time: float,
     p_mid_old = p_mid
     count = 0
 
-    times   = np.empty(max_events, np.float32)
-    p_mids  = np.empty(max_events, np.float32)
-    p_refs  = np.empty(max_events, np.float32)
+    times   = np.empty(max_events, np.float64)
+    p_mids  = np.empty(max_events, np.float64)
+    p_refs  = np.empty(max_events, np.float64)
     sides   = np.empty(max_events, np.int8)
     depths  = np.empty(max_events, np.int8)
     events  = np.empty(max_events, np.int8)
@@ -49,7 +49,7 @@ def simulate_QRM_jit(time: float,
     while True:
         # build rates
         n_rates = 2 * K * 3
-        rates   = np.empty((n_rates, 4), np.float32)
+        rates   = np.empty((n_rates, 4), np.float64)
         idx     = 0
         total   = 0.0
         for s in range(2):
